% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coated.R
\docType{data}
\name{Coated}
\alias{Coated}
\title{Coated}
\format{
A data frame with 15 rows and 5 columns. The columns
are as follows:
\describe{
  \item{Location}{The location numbers of observations.}
  \item{Coating1.Depth1}{The Depth values in the first treatment} 
  \item{Coating1.Number1}{The Number values in the first treatment}
  \item{Coating2.Depth2}{The Depth values in the second treatment} 
  \item{Coating2.Number2}{The Number values in the second treatment}
}
}
\source{
The data set is used in the book entitled
Methods of Multivariate Analysis (Rencher,2003).
}
\usage{
Coated
}
\description{
The data set is given in Table 5.3 in Rencher (2003).
The data set consists of 2 variables (Depth and Number), 2 treatments and 
15 observations.
The first column of the data is Location numbers.
}
\references{
Rencher, A. C. (2003). Methods of multivariate analysis 
(Vol. 492). John Wiley & Sons.
}
\keyword{datasets}
