% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWRcca.R
\name{.awrcca_make_psd}
\alias{.awrcca_make_psd}
\title{Ensure PSD correlation matrix (nearPD if needed)}
\usage{
.awrcca_make_psd(R, eps_psd = 1e-10)
}
\description{
Ensure PSD correlation matrix (nearPD if needed)
}
\keyword{internal}
