% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWRcca.R
\name{.awrcca_reweight_scores}
\alias{.awrcca_reweight_scores}
\title{Robust reweighting in score space using MCD distances}
\usage{
.awrcca_reweight_scores(S, alpha = 0.975)
}
\description{
Robust reweighting in score space using MCD distances
}
\keyword{internal}
