% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWRcca.R
\name{.awrcca_svd_cca}
\alias{.awrcca_svd_cca}
\title{Solve CCA via SVD from a regularized correlation matrix}
\usage{
.awrcca_svd_cca(Rreg, p, q)
}
\description{
Solve CCA via SVD from a regularized correlation matrix
}
\keyword{internal}
