% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWRcca.R
\name{.awrcca_weighted_cor}
\alias{.awrcca_weighted_cor}
\title{Weighted correlation matrix with 0/1 weights}
\usage{
.awrcca_weighted_cor(Z, w)
}
\description{
Weighted correlation matrix with 0/1 weights
}
\keyword{internal}
