% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWRcca.R
\name{.awrcca_wrap_block}
\alias{.awrcca_wrap_block}
\title{Wrap a data block columnwise after robust standardization}
\usage{
.awrcca_wrap_block(M, b, c, center_fun, scale_fun)
}
\description{
Wrap a data block columnwise after robust standardization
}
\keyword{internal}
