% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWRcca.R
\name{.awrcca_xi}
\alias{.awrcca_xi}
\title{Fisher-consistency correction factor xi under N(0,1)
xi = 1 / sqrt(E[psi(Z)^2])}
\usage{
.awrcca_xi(b, c, mc_n = 2e+05, seed = NULL)
}
\description{
Fisher-consistency correction factor xi under N(0,1)
xi = 1 / sqrt(E[psi(Z)^2])
}
\keyword{internal}
