% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iris.R
\docType{data}
\name{iris}
\alias{iris}
\title{Iris Data}
\format{
A data frame with 150 rows and 5 columns. The columns
are as follows:
\describe{
  \item{Sepal.Length}{The Sepal length values of iris flowers}
  \item{Sepal.Width}{The Sepal width values of iris flowers} 
  \item{Petal.Length}{The Petal length values of iris flowers}
  \item{Petal.Width}{The Petal width values of iris flowers} 
  \item{Species}{The species of iris flowers}
}
}
\source{
https://archive.ics.uci.edu/ml/datasets/Iris
}
\usage{
iris
}
\description{
The Iris dataset is consists of 4 variables, 3 groups and 150 observations.
The last column of the data is Iris species.
}
\keyword{datasets}
