% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simRobCellT2_onesample.R
\name{simRobCellT2_onesample}
\alias{simRobCellT2_onesample}
\title{Monte Carlo Simulation for d and q Constants of RobCellT2_onesample}
\usage{
simRobCellT2_onesample(
  n,
  p,
  nrep = 3000,
  alpha = 0.75,
  quant = 0.99,
  crit = 1e-04,
  seed = NULL,
  ...
)
}
\arguments{
\item{n}{The sample size.}

\item{p}{The number of variables.}

\item{nrep}{The number of Monte Carlo replications. Default is 3000.}

\item{alpha}{The cellMCD alpha parameter. Default is 0.75.}

\item{quant}{Quantile used in the cellMCD procedure. Default is 0.99.}

\item{crit}{Convergence criterion used in the cellMCD procedure. Default is 1e-04.}

\item{seed}{Optional random seed.}

\item{...}{Additional arguments passed to \code{cellWise::cellMCD()}.}
}
\value{
A list with the following elements:
\item{d}{The scaling constant of the approximate F distribution.}
\item{q}{The second degree of freedom of the approximate F distribution.}
\item{mean.T2}{The Monte Carlo mean of the simulated T^2 statistics.}
\item{var.T2}{The Monte Carlo variance of the simulated T^2 statistics.}
\item{n.success}{The number of successful Monte Carlo replications.}
}
\description{
Computes the constants \code{d} and \code{q} required for the approximate
F distribution of the cellMCD-based robust one-sample Hotelling T^2 statistic.
}
\examples{
if (requireNamespace("MASS", quietly = TRUE) &&
    requireNamespace("cellWise", quietly = TRUE)) {
  simRobCellT2_onesample(n = 50, p = 5, nrep = 50, seed = 123)
}

}
\author{
Hasan BULUT <hasan.bulut@omu.edu.tr>
}
